+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={3067--3086},
   year={2012},
   publisher={Optical Society of America}
+  doi={10.1364/boe.3.003067}
 }	
 
 @article{zhang2010real,
@@ -18,6 +19,7 @@
   pages={11772--11784},
   year={2010},
   publisher={Optical Society of America}
+  doi={10.1364/oe.18.011772}
 }
 
 @inproceedings{li2011scalable,
@@ -27,6 +29,7 @@
   pages={49--56},
   year={2011},
   organization={IEEE}
+  doi={10.1109/fccm.2011.27}
 }
 
 @article{wojtkowski2004ultrahigh,
@@ -38,6 +41,7 @@
   pages={2404--2422},
   year={2004},
   publisher={Optical Society of America}
+  doi={10.1364/OPEX.12.002404}
 }
 
 @article{cense2004ultrahigh,
@@ -49,6 +53,7 @@
   pages={2435--2447},
   year={2004},
   publisher={Optical Society of America}
+  doi={10.1364/OPEX.12.002435}
 }
 
 @article{yu2007improved,
@@ -60,6 +65,7 @@
   pages={7634--7641},
   year={2007},
   publisher={Optical Society of America}
+  doi={10.1364/oe.15.007634}
 }
 
 @article{shen2017improving,
@@ -71,6 +77,7 @@
   pages={4887--4918},
   year={2017},
   publisher={Optical Society of America}
+  doi={10.1364/BOE.8.004887}
 }
 
 @inproceedings{moiseev2012digital,
@@ -81,6 +88,7 @@
   pages={82132C},
   year={2012},
   organization={International Society for Optics and Photonics}
+  doi={10.1117/12.911297}
 }
 
 @article{park2003real,
@@ -92,6 +100,7 @@
   pages={782--793},
   year={2003},
   publisher={Optical Society of America}
+  doi={10.1364/OE.11.000782}
 }
 
 @article{jian2013graphics,
@@ -103,6 +112,7 @@
   pages={026002},
   year={2013},
   publisher={International Society for Optics and Photonics}
+  doi={10.1117/1.JBO.18.2.026002}
 }
 
 @article{drexler1999vivo,
@@ -114,6 +124,7 @@
   pages={1221--1223},
   year={1999},
   publisher={Optical Society of America}
+  doi={10.1364/OL.24.001221}
 }
 
 @incollection{de2008spectral,
@@ -123,6 +134,7 @@
   pages={147--175},
   year={2008},
   publisher={Springer}
+  doi={10.1007/978-3-540-77550-8_5}
 }
 
 @article{moon2010reference,
@@ -134,6 +146,7 @@
   pages={24395--24404},
   year={2010},
   publisher={Optical Society of America}
+  doi={10.1364/OE.18.024395}
 }
 
 @article{wieser2014high,
@@ -145,6 +158,7 @@
   pages={2963--2977},
   year={2014},
   publisher={Optical Society of America}
+  doi={10.1364/BOE.5.002963}
 }
 
 @article{kolb2019live,
@@ -155,6 +169,7 @@
   number={3},
   year={2019},
   publisher={Public Library of Science}
+  doi={10.1371/journal.pone.0213144}
 }
 
 @misc{pilia2018raycaster,
