+++ paper/paper.bib
@@ -22,6 +22,30 @@
   doi={10.1364/oe.18.011772}
 }
 
+@article{rasakanthan2011processing,
+  title={Processing and rendering of Fourier domain optical coherence tomography images at a line rate over 524 kHz using a graphics processing unit},
+  author={Rasakanthan, Janarthanan and Sugden, Kate and Tomlins, Peter H},
+  journal={Journal of biomedical optics},
+  volume={16},
+  number={2},
+  pages={020505},
+  year={2011},
+  publisher={International Society for Optics and Photonics}
+  doi={10.1117/1.3548153}
+}
+
+@article{sylwestrzak2012four,
+  title={Four-dimensional structural and Doppler optical coherence tomography imaging on graphics processing units},
+  author={Sylwestrzak, Marcin and Szlag, Daniel and Szkulmowski, Maciej and Gorczynska, Iwona M and Bukowska, Danuta and Wojtkowski, Maciej and Targowski, Piotr},
+  journal={Journal of biomedical optics},
+  volume={17},
+  number={10},
+  pages={100502},
+  year={2012},
+  publisher={International Society for Optics and Photonics}
+  doi={10.1117/1.JBO.17.10.100502}
+}
+
 @inproceedings{li2011scalable,
   title={Scalable, high performance Fourier domain optical coherence tomography: Why FPGAs and not GPGPUs},
   author={Li, Jian and Sarunic, Marinko V and Shannon, Lesley},
