+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={12},
   pages={3067--3086},
   year={2012},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/boe.3.003067}
 }	
 
@@ -18,7 +18,7 @@
   number={11},
   pages={11772--11784},
   year={2010},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/oe.18.011772}
 }
 
@@ -30,7 +30,7 @@
   number={2},
   pages={020505},
   year={2011},
-  publisher={International Society for Optics and Photonics}
+  publisher={International Society for Optics and Photonics},
   doi={10.1117/1.3548153}
 }
 
@@ -42,7 +42,7 @@
   number={10},
   pages={100502},
   year={2012},
-  publisher={International Society for Optics and Photonics}
+  publisher={International Society for Optics and Photonics},
   doi={10.1117/1.JBO.17.10.100502}
 }
 
@@ -52,7 +52,7 @@
   booktitle={2011 IEEE 19th Annual International Symposium on Field-Programmable Custom Computing Machines},
   pages={49--56},
   year={2011},
-  organization={IEEE}
+  organization={IEEE},
   doi={10.1109/fccm.2011.27}
 }
 
@@ -64,7 +64,7 @@
   number={11},
   pages={2404--2422},
   year={2004},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/OPEX.12.002404}
 }
 
@@ -76,7 +76,7 @@
   number={11},
   pages={2435--2447},
   year={2004},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/OPEX.12.002435}
 }
 
@@ -88,7 +88,7 @@
   number={12},
   pages={7634--7641},
   year={2007},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/oe.15.007634}
 }
 
@@ -100,7 +100,7 @@
   number={11},
   pages={4887--4918},
   year={2017},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/BOE.8.004887}
 }
 
@@ -111,7 +111,7 @@
   volume={8213},
   pages={82132C},
   year={2012},
-  organization={International Society for Optics and Photonics}
+  organization={International Society for Optics and Photonics},
   doi={10.1117/12.911297}
 }
 
@@ -123,7 +123,7 @@
   number={7},
   pages={782--793},
   year={2003},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/OE.11.000782}
 }
 
@@ -135,7 +135,7 @@
   number={2},
   pages={026002},
   year={2013},
-  publisher={International Society for Optics and Photonics}
+  publisher={International Society for Optics and Photonics},
   doi={10.1117/1.JBO.18.2.026002}
 }
 
@@ -147,7 +147,7 @@
   number={17},
   pages={1221--1223},
   year={1999},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
   doi={10.1364/OL.24.001221}
 }
 
@@ -157,7 +157,7 @@
   booktitle={Optical Coherence Tomography},
   pages={147--175},
   year={2008},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/978-3-540-77550-8_5}
 }
 
@@ -169,7 +169,7 @@
   number={24},
   pages={24395--24404},
   year={2010},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
[…]
