+++ paper/paper.md
@@ -92,7 +92,7 @@ The interactive volume viewer displays acquired OCT volumes without cropping or 
 In order to avoid unnecessary data transfer to host memory, CUDA-OpenGL interoperability is used which allows the processed data to remain in GPU memory for visualization. 
 However, it is possible to transfer the data to the host memory to save it on the hard disk, display individual axial depth profiles, so-called A-scans, in a 1D plot or use it within custom Extensions.
 
-Every processing step, except data conversion and IFFT, can be enabled and disabled during processing. To illustrate the effect of singe processing steps, B-scans of an OCT phantom (APL-OP01, Arden Photonics, UK) were acquired with a custom made SS-OCT system without k-klocking and with a slight dispersion imbalance. The acquired raw data was processed multiple times, each time with a different processing step disabled, see Fig. \ref{fig:effectofprocessing}. 
+Every processing step, except data conversion and IFFT, can be enabled and disabled during processing. To illustrate the effect of single processing steps, B-scans of an OCT phantom (APL-OP01, Arden Photonics, UK) were acquired with a custom made SS-OCT system without k-klocking and with a slight dispersion imbalance. The acquired raw data was processed multiple times, each time with a different processing step disabled, see Fig. \ref{fig:effectofprocessing}. 
 
  ![Effect of disabling single processing steps on resulting B-scan. The B-scans show a test pattern of an OCT phantom (APL-OP01, Arden Photonics, UK). Below each B-scan is an enlarged view of the corresponding area framed in red within the B-scan. a) The full processing pipeline, as described in section 3, is enabled. b) k linearization is disabled (all other steps are enabled). c) Dispersion compensation is disabled (all other steps are enabled). d) Windowing is disabled (all other steps are enabled). e) Fixed-pattern noise removal is disabled (all other steps are enabled). The red arrows point to horizontal structural artifacts that are visible if fixed-pattern noise removal is disabled. \label{fig:effectofprocessing}](figures/overview.png) 
 
@@ -116,7 +116,7 @@ Without 3D live view:| |
    Volumes per second|~**$2.2$**|~**$36$**
 
 
-Performance profiling for the GPU processing pipeline on the lab computer and with the same raw data set as used in table 1, was performed with NVIDIA Visual Profiler. A screenshot of it can be seen in figure \ref{fig:visualprofiler}, in which the relative duration of each kernel (function that is executed multiple times in parallel on the GPU) can be seen. It should be noted that the execution time of each kernel depends on the input data length and changes when different OCT data set dimensions are used. However, the k-linearization kernel, which also contains windowing and dispersion compensation, and the IFFT kernel need the most computing time.
+Performance profiling for the GPU processing pipeline on the lab computer and with the same raw data set as used in table 1, was performed with NVIDIA Visual Profiler. A screenshot of it can be seen in Figure \ref{fig:visualprofiler}, in which the relative duration of each kernel (function that is executed multiple times in parallel on the GPU) can be seen. It should be noted that the execution time of each kernel depends on the input data length and changes when different OCT data set dimensions are used. However, the k-linearization kernel, which also contains windowing and dispersion compensation, and the IFFT kernel need the most computing time.
 
   ![Screenshot of NVIDIA Visual Profiler showing timing sequence of the GPU processing pipeline for the same raw data set as used in table 1. Individual processing kernels are marked alphabetically: a) data conversion, b) kernel that combines k-linearization, windowing and dispersion compensation, c) IFFT, d) subtraction step of fixed pattern noise removal, e) truncate and logarithm, f) backward scan correction, g) copy B-scan frame to display buffer, h) copy en face view to display buffer \label{fig:visualprofiler}](figures/visualprofiler_screenshot.png)
 
