### 2020-06-08 #2245 @arfon
:wave: @spectralcode - thanks for your submission to JOSS. I'll be your editor for this submission. A couple of items I could do with your help with:

- Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
- The paper as currently written is rather long compared to our guidelines of 250--1000 words: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain . Please could you look at moving some of the content to documentation?
