### 2020-06-09 d672b4d
paper shortened and processing documentation extended
