+++ paper/paper.bib
@@ -56,77 +56,6 @@
   doi={10.1109/fccm.2011.27}
 }
 
-@article{wojtkowski2004ultrahigh,
-  title={Ultrahigh-resolution, high-speed, Fourier domain optical coherence tomography and methods for dispersion compensation},
-  author={Wojtkowski, Maciej and Srinivasan, Vivek J and Ko, Tony H and Fujimoto, James G and Kowalczyk, Andrzej and Duker, Jay S},
-  journal={Optics express},
-  volume={12},
-  number={11},
-  pages={2404--2422},
-  year={2004},
-  publisher={Optical Society of America},
-  doi={10.1364/OPEX.12.002404}
-}
-
-@article{cense2004ultrahigh,
-  title={Ultrahigh-resolution high-speed retinal imaging using spectral-domain optical coherence tomography},
-  author={Cense, Barry and Nassif, Nader A and Chen, Teresa C and Pierce, Mark C and Yun, Seok-Hyun and Park, B Hyle and Bouma, Brett E and Tearney, Guillermo J and De Boer, Johannes F},
-  journal={Optics express},
-  volume={12},
-  number={11},
-  pages={2435--2447},
-  year={2004},
-  publisher={Optical Society of America},
-  doi={10.1364/OPEX.12.002435}
-}
-
-@article{yu2007improved,
-  title={Improved lateral resolution in optical coherence tomography by digital focusing using two-dimensional numerical diffraction method},
-  author={Yu, Lingfeng and Rao, Bin and Zhang, Jun and Su, Jianping and Wang, Qiang and Guo, Shuguang and Chen, Zhongping},
-  journal={Optics express},
-  volume={15},
-  number={12},
-  pages={7634--7641},
-  year={2007},
-  publisher={Optical Society of America},
-  doi={10.1364/oe.15.007634}
-}
-
-@article{shen2017improving,
-  title={Improving lateral resolution and image quality of optical coherence tomography by the multi-frame superresolution technique for 3D tissue imaging},
-  author={Shen, Kai and Lu, Hui and Baig, Sarfaraz and Wang, Michael R},
-  journal={Biomedical optics express},
-  volume={8},
-  number={11},
-  pages={4887--4918},
-  year={2017},
-  publisher={Optical Society of America},
-  doi={10.1364/BOE.8.004887}
-}
-
-@inproceedings{moiseev2012digital,
-  title={Digital refocusing in optical coherence tomography},
-  author={Moiseev, Alexander A and Gelikonov, Grigory V and Shilyagin, Pavel A and Terpelov, Dmitry A and Gelikonov, Valentin M},
-  booktitle={Optical Coherence Tomography and Coherence Domain Optical Methods in Biomedicine XVI},
-  volume={8213},
-  pages={82132C},
-  year={2012},
-  organization={International Society for Optics and Photonics},
-  doi={10.1117/12.911297}
-}
-
-@article{park2003real,
-  title={Real-time multi-functional optical coherence tomography},
-  author={Park, B Hyle and Pierce, Mark C and Cense, Barry and De Boer, Johannes F},
-  journal={Optics express},
-  volume={11},
-  number={7},
-  pages={782--793},
-  year={2003},
-  publisher={Optical Society of America},
-  doi={10.1364/OE.11.000782}
-}
-
 @article{jian2013graphics,
   title={Graphics processing unit accelerated optical coherence tomography processing at megahertz axial scan rate and high resolution video rate volumetric rendering},
   author={Jian, Yifan and Wong, Kevin and Sarunic, Marinko V},
@@ -137,28 +66,6 @@
   year={2013},
[…]

[…]
