+++ paper/paper.md
@@ -48,7 +48,7 @@ The software can be separated into three parts: main application, development ki
 
 Raw data, i.e. acquired spectral fringe pattern, from the OCT system is transferred to RAM until enough data for a user-defined amount of cross-sectional images, so-called B-scans, is acquired. Via direct memory access (DMA) this raw data batch is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be stored or post processing steps are desired the processed OCT data can be transferred back to RAM with the use of DMA. An overview of the processing steps is depicted in Fig. \ref{fig:processing}.
 
- ![Processing pipeline of OCTproZ v1.1.0 Raw data from the OCT system is transferred to host RAM and via direct memory access (DMA) this raw data is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be saved on the hard disk, it can be transferred back to host RAM using DMA. \label{fig:processing}](figures/processing_pipeline.png) 
+ ![Processing pipeline of OCTproZ v1.1.0. Raw data from the OCT system is transferred to host RAM and via direct memory access (DMA) this raw data is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be saved on the hard disk, it can be transferred back to host RAM using DMA. \label{fig:processing}](figures/processing_pipeline.png) 
 
 A detailed description of each processing step can be found in the software repository. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster. [@pilia2018raycaster]
 In order to avoid unnecessary data transfer to host memory, CUDA-OpenGL interoperability is used which allows the processed data to remain in GPU memory for visualization. 
