### 2020-06-09 #2245 @spectralcode
@arfon thank you for your comment regarding the paper length. I have shortened the paper drastically. Please let me know if you still have any concerns.

Here are some potential reviewers from the list:
- mlxd
- mstimberg
- adam-m-jcbs
- RainierBarrett

A very suitable reviewer who is not on the list would be datenwolf. However, I don't know if he is willing to review.
