### 2020-06-12 #2245 @mstimberg
:wave: @arfon. I could review from the technical side, but I have only minimal domain knowledge and don't know other software used in this area. I also have a number of reviews lined up, so this is more of an "ok, if you don't find anyone else" rather than an enthusiastic yes :wink:  (the software itself looks pretty impressive though!)
