### 2020-06-12 #2245 @datenwolf
I'm okay to review it. However I must stress that any review must clearly indicate my financial interest in Optores. Other than that, I can do it.
