### 2020-06-13 #2245 @arfon
> I'm okay to review it. However I must stress that any review must clearly indicate my financial interest in Optores. Other than that, I can do it.

OK, thanks for letting me know @datenwolf. I'd rather not avoid this conflict it at all possible.
