### 2020-06-22 1241bf9
support for custom resampling curve added
