+++ octproz/src/octproz.cpp
@@ -284,11 +284,12 @@ void OCTproZ::initActionsAndDocks() {
 	this->actionRecord->setEnabled(false);
 }
 
-void OCTproZ::initGui(){
+void OCTproZ::initGui() {
 	ui->setupUi(this);
 	this->initActionsAndDocks(); //initActionsAndDocks() must be called before initMenu()
 	this->sidebar->init(this->actionStart, this->actionStop, this->actionRecord, this->actionSelectSystem, this->actionSystemSettings);
 	this->initMenu();
+	this->forceUpdateProcessingParams();
 
 	//Message Console connects
 	connect(this, &OCTproZ::info, this->console, &MessageConsole::slot_displayInfo);
@@ -308,7 +309,7 @@ void OCTproZ::initGui(){
 	qDebug() << "Main Thread ID: " << QThread::currentThreadId();
 }
 
-void OCTproZ::prepareDockWidget(QDockWidget*& dock, QWidget* widgetForDock, QAction*& action, const QIcon& icon, QString iconText){
+void OCTproZ::prepareDockWidget(QDockWidget*& dock, QWidget* widgetForDock, QAction*& action, const QIcon& icon, QString iconText) {
 	dock->setWidget(widgetForDock);
 	this->addDockWidget(Qt::RightDockWidgetArea, dock, Qt::Horizontal);
 	dock->setVisible(false);
@@ -321,8 +322,10 @@ void OCTproZ::prepareDockWidget(QDockWidget*& dock, QWidget* widgetForDock, QAct
 	this->viewToolBar->addAction(action);
 }
 
-void OCTproZ::initMenu(){
+void OCTproZ::initMenu() {
 	//file menu
+	this->actionSelectSystem->setShortcut(QKeySequence::Open);
+	this->actionSystemSettings->setShortcut(QKeySequence(Qt::CTRL + Qt::SHIFT + Qt::Key_O));
 	QMenu *fileMenu = menuBar()->addMenu(tr("&File"));
 	fileMenu->addAction(this->actionSelectSystem);
 	fileMenu->addAction(this->actionSystemSettings);
@@ -346,23 +349,45 @@ void OCTproZ::initMenu(){
 	//add toolbar view actions to view menu
 	viewMenu->addActions({ viewToolBarAction, view2DExtrasToolBarAction, controlToolBarAction, this->action1D, this->action2D, this->actionEnFaceView, this->action3D, this->actionConsole });
 
-	//settings menu
+	//extras menu
 	this->extrasMenu = this->menuBar()->addMenu(tr("&Extras"));
-//  QAction *settingsAct = this->extrasMenu->addAction(tr("&Processing settings"), this, &OCTproZ::slot_menuApplicationSettings);
-//  settingsAct->setStatusTip(tr("OCT processing settings"));
-//  settingsAct->setIcon(QIcon(":/icons/octproz_processingsettings_icon.png"));
+	QMenu* klinMenu = this->extrasMenu->addMenu(tr("&Resampling curve for k-linearization"));
+	klinMenu->setToolTipsVisible(true);
+	klinMenu->setStatusTip(tr("Settings for k-linearization resampling curve"));
+	klinMenu->setIcon(QIcon(":/icons/octproz_klincurve_icon.png"));
+	QActionGroup* customKlinCurveGroup = new QActionGroup(this);
+	this->actionSetCustomKLinCurve = new QAction(tr("&Load custom curve from file..."), this);
+	connect(this->actionSetCustomKLinCurve, &QAction::triggered, this, &OCTproZ::slot_loadCustomResamplingCurve);
+	this->actionUseSidebarKLinCurve = new QAction(tr("Use &polynomial curve from sidebar"), this);
[…]
