### 2020-06-22 d616d62
paper updated for v1.2.0
