+++ paper/paper.md
@@ -39,7 +39,7 @@ A commendable exemption is the GPU accelerated OCT processing pipeline published
 
 OCTproZ performs live signal processing and visualization of OCT data. It is written in C++, uses the cross-platform application framework Qt for the GUI and utilizes Nvidia’s computer unified device architecture (CUDA) for GPU parallel computing. A screenshot of the application can be seen in Fig. \ref{fig:screenshot}.
 
- ![Screenshot of OCTproZ v1.0. Processing settings visible in the left panel can be changed before processing is started or while processing is in progress. Processed data is live visualized in 2D as cross sectional images (B-scan and en face view) and in 3D as interactive volume rendering. The live view shows a piece of wood with a couple layers of tape and a laser burned hole. \label{fig:screenshot}](figures/20191122_screenshot3d.png)
+ ![Screenshot of OCTproZ v1.0.0 Processing settings visible in the left panel can be changed before processing is started or while processing is in progress. Processed data is live visualized in 2D as cross sectional images (B-scan and en face view) and in 3D as interactive volume rendering. The live view shows a piece of wood with a couple layers of tape and a laser burned hole. \label{fig:screenshot}](figures/20191122_screenshot3d.png)
 
 
 The software can be separated into three parts: main application, development kit (DevKit) and plug-ins. The main application, OCTproZ itself, contains the logic for GUI, processing and visualization. The DevKit, which is implemented as static library, provides the necessary interface for plugin development. Plug-ins can be one of two kinds: “Acquisition Systems” or “Extensions”. The former represent software implementations of physical or virtual OCT system hardware, the later are software modules that extend the functionality of an OCT system (e.g. software control of a liquid lens) or provides additional custom defined post processing steps. Both, Acquisition Systems and Extensions, are dynamic libraries that can be loaded into OCTproZ during runtime.  
@@ -48,13 +48,13 @@ The software can be separated into three parts: main application, development ki
 
 Raw data, i.e. acquired spectral fringe pattern, from the OCT system is transferred to RAM until enough data for a user-defined amount of cross-sectional images, so-called B-scans, is acquired. Via direct memory access (DMA) this raw data batch is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be stored or post processing steps are desired the processed OCT data can be transferred back to RAM with the use of DMA. An overview of the processing steps is depicted in Fig. \ref{fig:processing}.
 
- ![Processing pipeline of OCTproZ v1.1.0. Raw data from the OCT system is transferred to host RAM via direct memory access (DMA). This raw data is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be saved on the hard disk, it can be transferred back to host RAM using DMA. \label{fig:processing}](figures/processing_pipeline.png) 
+ ![Processing pipeline of OCTproZ v1.2.0. Raw data from the OCT system is transferred to host RAM via direct memory access (DMA). This raw data is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be saved on the hard disk, it can be transferred back to host RAM using DMA. \label{fig:processing}](figures/processing_pipeline.png) 
 
-A detailed description of each processing step can be found in the software repository. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster. [@pilia2018raycaster]
+A detailed description of each processing step can be found in the user manual. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster. [@pilia2018raycaster]
[…]
