### 2020-06-22 2f2b20f
updated version to 1.2.0
