### 2020-07-01 #2245 @spectralcode
Hi @marksgraham I can assist you to compile the code and then I will update the compiling instructions to help future users with similar troubles.

Can you provide some more information? Do you use the QtCreator and do you have installed cuda?
Are there any error messages from the compiler?
