### 2020-07-01 #2245 @marksgraham
I have installed QTCreator and I have all the CUDA drivers. OCTProZ_DevKit seems to build fine. virtualoctsystem gives three errors when I attempt to build, all related to `ceil`. Here is an example of one of them:

```
/home/mark/Downloads/OCTproZ-master/octproz_virtual_oct_system/src/virtualoctsystem.cpp:60: error: ‘ceil’ was not declared in this scope
  size_t bufferSize = currParams.width*currParams.height*currParams.depth*ceil((double)this->currParams.bitDepth / 8.0);
                                                                          ^~~~
````
