### 2020-07-01 #2245 @spectralcode
@marksgraham thank you for taking the time to try to compile this somewhat large project. I'm sure we can get it to work together.

I've updated virtualoctsystem.h (just added the line "#include "math.h"").
Could you please retry it with the new virtualoctsystem.h and let me know if it works?

If this does not work another possibility is to replace "ceil" by "std::ceil", but as this problem does not occur with Ubuntu 16.04 I will first have to install Ubuntu 18.04 and test it.
