### 2020-07-01 #2245 @marksgraham
@spectralcode I can now compile virtualoctsystem. It isn't clear to me how I actually get it to run now.. when I press run I'm told `You need to set an executable in the custom run configuration.`
