### 2020-07-01 #2245 @spectralcode
@marksgraham ok, thats great!
The virtual OCT system is an OCTproZ plugin and it cannot be executed on its own. It needs to be loaded from OCTproZ. Have a look at the quick start guide in the user manual: https://spectralcode.github.io/OCTproZ/index.html

Try compiling OCTproZ and then run it to load the virtual OCT system.
(Hint: before running OCTproZ with the "play" button in Qt Creator right click on the OCTproZ project and click on the first option. It should be something like "set as active project")
