### 2020-07-01 #2245 @marksgraham
@spectralcode
When I now try to compile OCTproZ I get a CUDA error:
/bin/sh: 1: /usr/local/cuda/bin/nvcc: not found
I had  thought I had the relevant drivers because nvidia-smi works on my system with the following output:
``` NVIDIA-SMI 430.50       Driver Version: 430.50       CUDA Version: 10.1   ```

Does this mean I also need to install the nvidia-cuda-toolkit to get this working on my system?
