### 2020-07-02 #2245 @marksgraham
I have nvcc installed at `/usr/bin/nvcc` but QTcreator is still looking for it in `/usr/local/cuda/bin/nvcc`. It isn't  clear to me how I can configure the build to point it to the right place - any ideas?
