### 2020-07-02 #2245 @arfon
@datenwolf - could you say a little more about your financial interests in Optores here? It's not clear what _Optores_ is, and how it's related to this software.
