### 2020-07-02 #2245 @spectralcode
@marksgraham I am not sure, but it could be that you need to do the "post-installation actions" which are listed in the [official cuda installation guide for Linux](https://docs.nvidia.com/cuda/cuda-installation-guide-linux).

I am going to set up Ubuntu 18.04 and the Qt + Cuda development environment to see if I run in to the same issue.
