### 2020-07-02 #2245 @spectralcode
I've tested it with Ubuntu 18.04 and everything worked just fine. Unfortunately (?) I did not had the same problem as you did with nvcc that can't be found.
For installation of cuda I followed the official cuda installation guide for Linux. If you have also used this guide, then probably something went wrong during the post-installation actions as those include the setup of the PATH variable.
Do you think this can be the case?

But I've found some other minor issues in the .pro files that are now fixed.
