### 2020-07-03 0a6d2ac
instructions for setting up the development environment added
