### 2020-07-03 df88eef
fixed bug that occurs on Linux systems that does not allow the virtual oct system to stop properly
