### 2020-07-03 #2245 @marksgraham
Hi @spectralcode @arfon

I've delved deeper and concluded that for me to get this working I would need to do a fresh install of all my CUDA drivers. Experience had taught me that messing with CUDA drivers can cause various problems and should only be done when absolutely necessary. I'm reluctant to do so in order to complete this review, especially as I'm doing a lot of GPU-heavy work right now that would be interrupted if I were to interfere with my setup. So I'm going to have to say sorry, and pass on this.
