### 2020-07-07 ce258bb
zoom and range mouse interaction for yAxis of 1d raw plot added
