+++ paper/paper.md
@@ -48,9 +48,9 @@ The software can be separated into three parts: main application, development ki
 
 Raw data, i.e. acquired spectral fringe pattern, from the OCT system is transferred to RAM until enough data for a user-defined amount of cross-sectional images, so-called B-scans, is acquired. Via direct memory access (DMA) this raw data batch is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be stored or post processing steps are desired the processed OCT data can be transferred back to RAM with the use of DMA. An overview of the processing steps is depicted in Fig. \ref{fig:processing}.
 
- ![Processing pipeline of OCTproZ v1.2.0. Raw data from the OCT system is transferred to host RAM via direct memory access (DMA). This raw data is then copied asynchronously to GPU memory where OCT signal processing is executed. If the processed data needs to be saved on the hard disk, it can be transferred back to host RAM using DMA. \label{fig:processing}](figures/processing_pipeline.png) 
+ ![Processing pipeline of OCTproZ v1.2.0. Each box inside "OCTproZ GPU Processing" represents a CUDA kernel. Some processing steps are combinend into a single kernel (e.g. k-linearization, dispersion compensation and windowing) to enhance processing performance. \label{fig:processing}](figures/processing_pipeline_linear_v1_1_0.png) 
 
-A detailed description of each processing step can be found in the user manual. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster. [@pilia2018raycaster]
+The processing pipeline consists of data conversion, k-linearization, numerical dispersion compensation, windowing, fast Fourier transform, fixed-pattern noise removal, truncate, logarithm, backward scan correction, sinusoidal scan correction and visualization. A detailed description of each processing step can be found in the user manual. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster. [@pilia2018raycaster]
 In order to avoid unnecessary data transfer to host memory, CUDA-OpenGL interoperability is used which allows the processed data to remain in GPU memory for visualization. 
 
 # 4. Processing Performance 
@@ -77,7 +77,7 @@ Without 3D live view:| |
 
 To develop custom plug-ins for OCTproZ and thus extends its functionality, a development kit is provided. It consists of a static library and a collection of C++ header files that specify which classes and methods have to be implemented to create custom plug-ins. Currently two kinds of plug-ins exist: Acquisition Systems and Extensions. For both we made examples including source code publicly available which may be used together with the open source and cross-platform integrated development environment Qt Creator as starting point for custom plug-in development.
 
-For Acquisition System development the two key methods that need to be implemented are “startAcquisition()” and “stopAcquisition()”. In startAcquisition() an acquisition buffer needs to be filled and the corresponding boolean flag needs to be set. The processing thread in the main application continuously checks the acquisition buffer flag to transfer the acquired raw data to GPU as soon as the acquisition buffer is filled. When the acquisition is stopped, stopAcquisition() is called, where termination commands to stop hardware such as scanners may be implemented. 
+An Acquisition System represents an OCT Systems and its main task is to provide raw data to OCTproZ. In most cases, this means that the implementation of an Acquisition System contains the software control of a data acquisition unit.
 
[…]
