+++ paper/paper.md
@@ -77,7 +77,7 @@ Without 3D live view:| |
 
 To develop custom plug-ins for OCTproZ and thus extends its functionality, a development kit is provided. It consists of a static library and a collection of C++ header files that specify which classes and methods have to be implemented to create custom plug-ins. Currently two kinds of plug-ins exist: Acquisition Systems and Extensions. For both we made examples including source code publicly available which may be used together with the open source and cross-platform integrated development environment Qt Creator as starting point for custom plug-in development.
 
-An Acquisition System represents an OCT Systems and its main task is to provide raw data to OCTproZ. In most cases, this means that the implementation of an Acquisition System contains the software control of a data acquisition unit.
+An Acquisition System represents an OCT System and its main task is to provide raw data to OCTproZ. In most cases, this means that the implementation of an Acquisition System contains the software control of a data acquisition unit.
 
 Extensions have a wide area of use cases. As they are able to receive raw data and processed data via the Qt signals and slots mechanism, they are suitable for custom post-processing routines. The exact implementation of an Extension is mainly up to the developer and can also include hardware control. Therefore, Extensions are ideal for hardware control algorithms that rely on feedback from live OCT images. The best example of this is wavefront sensorless adaptive optics with a wavefront modulator such as a deformable mirror.  Particular care must be taken if high speed OCT systems are used, as the acquisition speed of OCT data may exceed the processing speed of the custom Extension. In this case, a routine within the Extension should be implemented that discards incoming OCT data if previous data is still processed. 
 
