### 2020-07-10 #2245 @spectralcode
@arfon - I still hope that @datenwolf will reply.

Here are some other possible candidates:
- Marcin Sylwestrzak (http://lab.fizyka.umk.pl/team/ )
- Pete Tomlins (https://www.qmul.ac.uk/dentistry/people/profiles/drpetetomlins.html )
- Christian Viehland (http://biophotonics.pratt.duke.edu/content/viehland ) or one of the Phd students (e. g. David Li http://biophotonics.pratt.duke.edu/content/li )
- GitHub user PhillyVanilly9119 (I talked with him about the Software and he starred the Project on GitHub, so I don't know if this disqualifies him as reviewer.)
