### 2020-07-15 1e175b6
fixed a bug that occured on Linux systems and made it impossible to right click on a plot in the sidebar and open a file dialog.
