### 2020-07-21 939eec8
fixed bug that made it impossible to open custom resampling curve on Ubuntu
