### 2020-08-04 #2245 @arfon
:wave: @PhillyVanilly9119 - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
