### 2020-08-05 #2245 @PhillyVanilly9119
Hey guys!
First of all, thank you guys for considering me as a possible reviewer (@arfon ). However, I think I might be too biased, since my research is on a very related topic (as @spectralcode mentioned). I like to think I'm not but in the interest of fact based reviewing, I agree that someone else would proably be a more suitable candidate.
I think guys from Jo Izatt group, like Christian Viehland or his PhD students working on intra-surgical OCT would be a great fit!
@spectralcode : I think they are very active in reviewing papers in OSA jourals, so I think you can directly contact them and you dint have to wait for the tweet to appear on octnews.org.

Good luck with your search!
Cheers
