### 2020-08-05 #2245 @arfon
> @spectralcode : I think they are very active in reviewing papers in OSA jourals, so I think you can directly contact them and you don't have to wait for the tweet to appear on octnews.org.

:zap: thanks @PhillyVanilly9119! If you are willing to contact them that would be great! Feel free to copy me (arfon.smith@gmail.com) on the email thread.
