### 2020-08-05 #2245 @spectralcode
@PhillyVanilly9119 : thanks!
Regarding octnews: A link to Arfon's tweet already was on [octnews](http://www.octnews.org/articles/9866790/the-journal-of-open-source-software-is-looking-for/) but it doesn't seem to have helped much.

@arfon : I think you have already contacted Christian Viehland (see [your post](https://github.com/openjournals/joss-reviews/issues/2245#issuecomment-658701574) ) but probably not David Li. So I just contacted David Li via mail.
