### 2020-08-06 88a62a1
readme updated - added performance info for activated 3D live view
