### 2020-08-06 f405230
readme updated, typo fixed
