### 2020-08-10 3fc780c
Order of the connections for (re-)activation of extensions changed. Info and error messages that occur during activation will now be also displayed in message console
