### 2020-08-12 10a7bc7
funding section in paper updated
