### 2020-08-17 #2245 @arfon
@phtomlins - while we're waiting to find a second reviewer, feel free to familiarize yourself with the JOSS reviewer guidelines https://joss.readthedocs.io/en/latest/reviewer_guidelines.html . Once we have a second reviewer, I'll go ahead and set up the actual review.
