### 2020-08-17 #2580 @whedon
[REVIEW]: Open Source Optical Coherence Tomography Software
**Submitting author:** @spectralcode (<a href="http://orcid.org/0000-0002-4494-6127">Miroslav Zabic</a>)
**Repository:** <a href="https://github.com/spectralcode/OCTproZ" target ="_blank">https://github.com/spectralcode/OCTproZ</a>
**Version:** v1.0.0
