### 2020-08-17 ce2495e
contribution guidelines updated
