+++ CONTRIBUTING.md
@@ -1,0 +1,29 @@
+ # <img style="vertical-align:middle" img src="images/octproz_icon.png" width="64"> OCTproZ - Contribution Guidlines
+
+This document provides guidelines on how to contribute to OCTproZ. 
+
+There are four main ways to contribute: issue reports, feature requests, code contributions and plug-in development.  
+
+
+Issue Reports
+----------
+- Check if the issue you want to report or a similar one was already reported on the [OCTproZ issue page](https://github.com/spectralcode/OCTproZ/issues). Feel free to comment on already existing issue reports to provide valuable information. 
+- Report new issues by using the "Bug report"-template if applicable.
+
+
+Feature Requests
+----------
+- Check if the feature you want to request or a similar one was already requested on the [OCTproZ issue page](https://github.com/spectralcode/OCTproZ/issues). Feel free to comment on already existing feature requests to provide valuable information. 
+- Request new features by using the "Feature request"-template if applicable.
+
+
+Code Contributions
+----------
+- Before creating a [pull request](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/about-pull-requests), please announce your plans in the [OCTproZ issue page](https://github.com/spectralcode/OCTproZ/issues). This enables an open discussion which will get the best out of your efforts. 
+- Currently there is no style guide. New code should be more or less consistent with the style of already existing code.
+
+
+Plug-ins
+----------
+- If you have developed a plug-in for OCTproZ and want it to be mentioned in the [OCTproZ README.md](https://github.com/spectralcode/OCTproZ/blob/master/README.md), please contact us.
+- Contact information can be found at the bottom of the [user manual](https://spectralcode.github.io/OCTproZ/index.html)
