### 2020-08-17 #2245 @brandondube
👋 @arfon @spectralcode I can perform this review.  I do not think it constitutes COI, but I am an alumni ('18) of a group that does research on OCT, e.g. [a recent publication](https://www.osapublishing.org/oe/abstract.cfm?uri=oe-28-14-19937&origin=search).  I never worked in any depth on any of the OCT stuff.
