### 2020-08-18 #2580 @brandondube
:wave: @spectralcode good news at the start - the 1.2 release opens on my desktop with no issue.  I did not easily find an example file in the repository or referenced in the documentation.  Is there one available to demo the software with?
