### 2020-08-18 #2580 @spectralcode
Hi @brandondube thank you very much for reviewing!
That's great that the software opens without issues. To get a link to a test data set, please have a look at the "Download and Installation" section in the readme.md. The same link can also be found in the quick start guide of the user manual.
