+++ vision.md
@@ -1,0 +1,20 @@
+ # <img style="vertical-align:middle" img src="images/octproz_icon.png" width="64"> OCTproZ - Vision and long term goals
+
+This document describes the long-term goals of OCTproZ.   
+
+
+General vision
+----------
+OCTproZ should:
+- be an easy to use OCT signal processing and visualization software for researchers that utilize OCT.
+- enable everyone to use custom build OCT systems without the hassle of having to write the processing software
+- __provide open OCT processing algorithms to make scientific research more reproducible__
+
+
+
+Long-term goals
+----------
+- Implement OCT processing pipeline in other frameworks besides CUDA (e.g. OpenCL, OpenMP, C ++ AMP) to support various computer hardware for processing
+- Multi GPU support
+- Plugin database for most common OCT system hardware configurations
+- Advanced visualization for live view of acquired OCT data (e.g. speckle variance, spectroscopic)
