### 2020-08-23 ae1e334
added developer guide
