### 2020-08-23 ef64fac
references in paper updated
