+++ paper/paper.bib
@@ -100,3 +100,16 @@
   publisher={​GitHub},
   url={https://github.com/m-pilia/volume-raycasting}
 }
+
+@misc{qt,
+	title = {Qt},
+	url = {https://www.qt.io},
+	urldate = {2020-08-23},
+}
+
+@misc{cuda,
+	title = {CUDA},
+  author={​NVIDIA},
+	url = {https://developer.nvidia.com/cuda-zone},
+	urldate = {2020-08-23},
+}
