+++ paper/paper.bib
@@ -102,14 +102,16 @@
 }
 
 @misc{qt,
-	title = {Qt},
-	url = {https://www.qt.io},
-	urldate = {2020-08-23},
+	title={Qt},
+  year={1995}
+	url={https://www.qt.io},
+	urldate={2020-08-23},
 }
 
 @misc{cuda,
-	title = {CUDA},
+	title={CUDA},
   author={​NVIDIA},
-	url = {https://developer.nvidia.com/cuda-zone},
-	urldate = {2020-08-23},
+  year={2007},
+	url={https://developer.nvidia.com/cuda-zone},
+	urldate={2020-08-23},
 }
