+++ paper/paper.bib
@@ -103,7 +103,7 @@
 
 @misc{qt,
 	title={Qt},
-  year={1995}
+  year={1995},
 	url={https://www.qt.io},
 	urldate={2020-08-23},
 }
