### 2020-08-24 #2580 @spectralcode
@brandondube thank you very much for your feedback!
**Performance:**
Thank you for testing the performance and providing this information!
Have you used the same data dimensions and the same batch size as stated in the paper? (1024 samples per raw A-scan, 512 A-scans per B-scan, 256 B-scans per volume and batch size: 256 B-scans per buffer)

**Paper:**
> Last line of section 5 should be "being processed."

&rarr; Fixed!

> You do not have one for Qt or Cuda. [...] please cite them.

&rarr; Done! I couldn't find any other JOSS paper citing Qt or CUDA so I am not sure if the way i cited it is the common way to do it. Please let me know if you would have done it differently.

> Do you have permission to use the EU and regional development fund logos?

&rarr; Thank you very much for pointing this out. I am even obliged to use the logo. In addition, the logo must be at least as large as the largest other logo. So maybe I have to enlarge it as the JOSS logo is quite big. I will talk with the administration regarding the logo size and maybe there is a way out not to use the logos at all as it seems quite unusual to me.

> For example, implementation with python and CuPy probably would have a much lower barrier to entry for scientists.

&rarr; Yes, I think you are right about that! There is an idea for the long run to extend the plugin system such that it enables scripting with Python, but no evaluations have been made yet in this regard. At the start of the project it did not cross my mind to use Python for a full blown desktop application that has specific requirements on memory management, processing speed and should be able to control all kinds of hardware. I knew all of these requirements could be met by C++ and CUDA, so I went for it.

**Software / Docs:**

>  I feel that not having a developer guide or similar is a strong impediment to success in that area.

&rarr;  This is a great comment. Thank you for pointing this out! I feel the same way, so [here](https://spectralcode.github.io/OCTproZ/developer.html) is the developer guide.

> The example dataset could be more boldly pointed out (I missed it the first time).

&rarr; Done! It now has its own section.

> I also did not find a test suite.

&rarr;  There is none. I hoped that providing test data and a detailed step by step guide how to use it would be sufficient. See the [JOSS review criteria for Tests](https://joss.readthedocs.io/en/latest/review_criteria.html?#tests).
