### 2020-08-27 #2580 @spectralcode
Just a small update: I received feedback regarding the logo and it should stay the way it is.
