### 2020-08-28 #2580 @brandondube
I do not remember the precise settings I used for the sample data configuration 😅  - been too many days.  I used the settings listed in the documentation where it is referenced.  If those are different to the settings in the paper, then the perf may differ.

No sweat on the logos - if you are required to use them by your funding agencies, that is quite a strong "permission to use."  We would not want to be in a situation where JOSS receives a C&D or similar on your paper over a logo; glad it's sorted.

The developer guide looks solid.

Regarding tests - I had in mind something like a GUI automation script that can verify the software correctly performs some action.  Your software meets the standard for JOSS, but there is nothing wrong with exceeding the standard :)

LGTM

-Brandon
