### 2020-08-28 #2580 @arfon
:wave: @jrasakanthan - thanks for agreeing to take the place of @phtomlins as a reviewer here. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
