### 2020-08-28 #2580 @jdavidli
@spectralcode
I've been doing testing on a computer with the following specs:
i9-9900X
RTX 2080Ti
64 GB RAM
Windows 10 Pro

I got similar performance numbers to @brandondube
Without 3D
3.7505 volumes/second
60 buffers/second
960 B-scans/second
A-scans/second 491520
Buffer size [MB] 26
Data throughput [MB/s] 1560

With 3D
3.7505 volumes/second
60 buffers/second
960 B-scans/second
A-scans/second 491520
Buffer size [MB] 26
Data throughput [MB/s] 1560

I did notice that the settings described in the documentation differs from what's described in the paper. The performance numbers above are using the settings from the documentation. When I tried using the settings in the paper, the application seems to hang for a few seconds and then just completely crashes.
