### 2020-08-28 #2580 @spectralcode
@jdavidli thank you very much for testing! I'm not sure why your are not able to replicate the processing rates from the paper. Here is some information that may help to clear things up:
- The settings described in the documentation are chosen such that the test data works on almost all GPUs.
- The settings described in the paper are optimized for the used GPU.
- The buffer size has major effect on processing speed. If it is to small the processing may be slower than possible. If it is too large the application may crash.
- A larger buffer size results in higher GPU memory usage, which can exceed the available memory on the used GPU (this usually causes the application to crash)

In the paper i used 1024 samples per raw A-scan as this is a more common value than 1664.
With the provided test data set you could also use 1024 samples per raw A-scan to verify the processing rates described in the paper but the resulting OCT images will look distorted of course. By the way, the provided test data set is exactly the same one I used in the paper.

With a RTX 2080Ti you should totally be able to use the same parameters as used for the lab system in the paper.
The crash you described puzzles me and I'm not sure what caused it.
One thing we could test is to increase the buffer size just slightly to see if processing speed increases or if the application still crashes: 1664 samples per raw A-scan, 512 A-scans per B-scan, 8 B-scans per buffer, 32 Buffers per volume, 32 Buffers to read from file
Could you retry it with these parameters, please?
