### 2020-08-28 #2580 @spectralcode
@jdavidli sorry, the parameters in my last message are incorrect. They actually reduce the buffer size!
Here the correct parameters: 1664 samples per raw A-scan, 512 A-scans per B-scan, 32 B-scans per buffer, 8 Buffers per volume, 8 Buffers to read from file
With those parameters the buffer size should be 52 MB
