### 2020-08-29 #2580 @jdavidli
> @jdavidli sorry, the parameters in my last message are incorrect. They actually reduce the buffer size!
> Here the correct parameters: 1664 samples per raw A-scan, 512 A-scans per B-scan, 32 B-scans per buffer, 8 Buffers per volume, 8 Buffers to read from file
> With those parameters the buffer size should be 52 MB

@spectralcode with your new suggested parameters, I get the following:
Volumes/second: 7.5
Buffers/second: 60
B-scans/second: 1920
A-scans/second: 9.8e5
Buffer size: 52
Data throughput: 3120

The volume rate is still significantly lower than what I would expect compared to your benchmarks in the paper. The A-scans/second also seems really low and I suspect might be might affecting the volume rate? Furthermore, with these settings, it seems like the test data set is no longer playing correctly. The B-scan does not change at all (unless I have it display another location within the volume).

I think the additional information you provided about parameter settings in your other comment is useful and would be good to include in either the paper or the manual. The labels in the system settings also differ slightly from the paper (e.g. bit depth vs bits per sample, B-scans per buffer and buffers per volume vs frames per volume). It would probably be a little bit more user-friendly if the language used was more consistent.

The included developer guide still feels a bit light. The examples you provided of extension plug-ins are pretty good. But looking through the source code of the virtual OCT system example for acquisition plug-ins, I feel like it would still be a bit challenging for others to use this software with their own systems and contribute to the project. If you have an example of an implementation with a physical OCT system rather than a virtual one, I think that would be incredibly useful to include as an example acquisition plug-in.
