### 2020-08-30 95466cf
detailed performance information added to documentation
