+++ performance.md
@@ -1,0 +1,83 @@
+ # <img style="vertical-align:middle" img src="images/octproz_icon.png" width="64"> OCTproZ - Performance Information
+
+Processing rate highly depends on the size of the raw data, the used computer hardware and resource usage by background or system processes. With modern computer hardware and typical data dimensions for OCT, OCTproZ achieves A-scan rates in the MHz range.
+
+A test data set with 12 bit per sample, 1024 samples per line, 512 lines per frame and 256 frames per volume was used to measure the performance on different systems:
+
+| |**Office Computer**|**Lab Computer**|**Gaming Computer**|
+|:-----|:-----|:-----|:-----|
+|CPU|Intel® Core i5-7500|AMD Ryzen™ Threadripper 1900X|AMD Ryzen™ 5 1600|
+|RAM|16 GB|32 GB|16 GB|
+|GPU|NVIDIA Quadro K620|NVIDIA GeForce GTX 1080 Ti| NVIDIA GeForce GTX 1080
+|Operating system|Windows 10|Ubuntu 16.04| Windows 10|
+|A-scan rate with 3D view| ~ 250 kHz ( ~1.9 volumes/s)|~ 4.0 MHz (~ 30 volumes/s)|~ 2.1 MHz (~ 16 volumes/s)|
+|A-scan rate without 3D view| ~ 300 kHz ( ~2.2 volumes/s)|~ 4.8 MHz (~ 36 volumes/s)|~ 2.7 MHz (~ 19 volumes/s)|
+
+Office Computer, Lab Computer: <br>
+The performance was measured with the full processing pipeline of OCTproZ v1.0.0. The same performance is expected with OCTproZ v1.2.0 if live sinusoidal scan distortion correction is disabled.
+
+Gaming Computer:<br>
+The performance was measured with OCTproZ v1.2.0 with disabled live sinusoidal scan distortion correction.
+
+Here are the relevant parameters that were used with Virtual OCT System and OCTproZ to determine the performance:
+
+| |**Office Computer**|**Lab Computer**|**Gaming Computer**|
+|:-----|:-----|:-----|:-----|
+|**Virtual OCT System Settings**| | |
+|bit depth [bits]|12|12|12|
+|Samples per raw A-scan|1024|1024|1024|
+|A-scan per B-scan|512|512|512|
+|B-scans per Buffer|32|256|256|
+|Buffers per volume|8|1|1|
+|Buffers to read from file|16|2|2|
+|Wait after file read [us]|100|100|100|
+|**OCTproZ Settings**| | |
+|Bit shift sample values by 4|enabled|enabled|enabled|
+|Flip every second B-scan|||
+|k-linearization|enabled|enabled|enabled|
+|Dispersion Compensation|enabled|enabled|enabled|
+|Windowing|enabled|enabled|enabled|
+|Fixed-Pattern Noise Removal|enabled|enabled|enabled|
+|B-scans for noise determination:|1|26|1|
+|&emsp; once at start of measurement|enabled|enabled|enabled|
+|&emsp;continuously|disabled|disabled|disabled|
+|Sinusoidal scan correction|disabled|disabled|disabled|
+|Log scaling|enabled|enabled|enabled|
+|Stream Processed Data to Ram|enabled|disabled|disabled|
+
+
+
+How to Determine Performance
+--------
+ OCTproZ provides live performance information within the sidebar in the "Processing"-tab. Live performance estimation is performed and updated every 5 seconds.
+ 
+ It is also possible to use the [NVIDIA Visual Profiler](https://developer.nvidia.com/nvidia-visual-profiler) to analyze performance in more detail.
+
[…]
