### 2020-08-30 #2580 @spectralcode
@jdavidli thank you very much, this is great that you was able to change the parameters and perform processing without crash. This tells me at least that there is not a major bug in the software that I have missed and that performance on your system can be increased by adjusting the Virtual OCT System parameters.

The documentation now contains [detailed information about processing performance](https://github.com/spectralcode/OCTproZ/blob/master/performance.md). Please let me know if this is helps you to replicate my measurements.

>The B-scan does not change at all

→ Yes, this is expected when _Buffers per volume_ and _Buffers to read from file_ have the same value. If you want the live view to change while using Virtual OCT System, double the value for _Buffers to read from file_.

>The labels in the system settings also differ slightly from the paper

→ Thank you very much for pointing this out! I changed it!

>If you have an example of an implementation with a physical OCT system rather than a virtual one, I think that would be incredibly useful to include as an example acquisition plug-in.

→ Yes, I totally agree with you! Unfortunately I cannot provide an implementation for actual OCT hardware at the moment.
