### 2020-08-30 483f01e
typo fixed in performance.md
