### 2020-08-30 cc82491
typo fixed in performance.md
