+++ performance.md
@@ -2,7 +2,7 @@
 
 Processing rate highly depends on the size of the raw data, the used computer hardware and resource usage by background or system processes. With modern computer hardware and typical data dimensions for OCT, OCTproZ achieves A-scan rates in the MHz range.
 
-A test data set with 12 bit per sample, 1024 samples per line, 512 lines per frame and 256 frames per volume was used to measure the performance on different systems:
+A test data set with 12 bit per sample, 1024 samples per raw A-scan, 512 A-scans per B-scan and 256 B-scans per volume was used to measure the performance on different systems:
 
 | |**Office Computer**|**Lab Computer**|**Gaming Computer**|
 |:-----|:-----|:-----|:-----|
