### 2020-08-30 33a113c
typos fixed in performance.md
