### 2020-08-30 fbe16c1
typo fixed in performance.md
