### 2020-08-30 e295fce
updated BUILD.md
