### 2020-09-02 12623db
citation style removed from paper
