+++ paper/paper.bib
@@ -102,8 +102,8 @@
 }
 
 @misc{qt,
-	title={Qt},
-  year={1995},
+	author={Qt},
+  year={2020},
 	url={https://www.qt.io},
 	urldate={2020-08-23},
 }
@@ -111,7 +111,7 @@
 @misc{cuda,
 	title={CUDA},
   author={​NVIDIA},
-  year={2007},
+  year={2020},
 	url={https://developer.nvidia.com/cuda-zone},
 	urldate={2020-08-23},
 }
