### 2020-09-02 23b9b14
performance information for gaming pc corrected and bar chart added
