### 2020-09-02 e6dcc2a
performance.md updated
