### 2020-09-04 a79949d
vision.md updated
