+++ vision.md
@@ -13,7 +13,6 @@ OCTproZ should:
 Long-term goals
 ----------
 - Implement OCT processing pipeline in other frameworks besides CUDA (e.g. OpenCL, OpenMP, C ++ AMP) to support various computer hardware for processing
-- Multi GPU support
 - Plugin database for most common OCT system hardware configurations
 - Advanced visualization for live view of acquired OCT data (e.g. speckle variance, spectroscopic)
 - Python or scripting interface either as plugin or integrated within OCTproZ
