### 2020-09-04 baf6945
vision.md updated
