### 2020-09-09 c7a9703
video link added to readme.md
