+++ octproz_project/octproz/src/settings.h
@@ -1,0 +1,204 @@
+/**
+**  This file is part of OCTproZ.
+**  OCTproZ is an open source software for processig of optical
+**  coherence tomography (OCT) raw data.
+**  Copyright (C) 2019-2020 Miroslav Zabic
+**
+**  OCTproZ is free software: you can redistribute it and/or modify
+**  it under the terms of the GNU General Public License as published by
+**  the Free Software Foundation, either version 3 of the License, or
+**  (at your option) any later version.
+**
+**  This program is distributed in the hope that it will be useful,
+**  but WITHOUT ANY WARRANTY; without even the implied warranty of
+**  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+**  GNU General Public License for more details.
+**
+**  You should have received a copy of the GNU General Public License
+**  along with this program. If not, see http://www.gnu.org/licenses/.
+**
+****
+** Author:	Miroslav Zabic
+** Contact:	zabic
+**			at
+**			iqo.uni-hannover.de
+****
+**/
+
+//!	Settings
+/*!	Class for storing processing, visualization and record settings to the hard drive.
+*/
+
+#ifndef SETTINGS_H
+#define SETTINGS_H
+
+#define SETTINGS_FILE_NAME "settings.ini"
+#define SETTINGS_PATH QStandardPaths::writableLocation(QStandardPaths::ConfigLocation) + "/" + SETTINGS_FILE_NAME
+#define TIMESTAMP "timestamp"
+#define REC "record"
+#define PROC "processing"
+#define VIZ "visualization"
+#define STREAM "streaming"
+#define MAIN "main_window_settings"
+#define MAIN_GEOMETRY "main_geometry"
+#define MAIN_STATE "main_state"
+#define REC_PATH "path"
+#define REC_MODE "record_mode"
+#define REC_STOP "stop_after_record"
+#define REC_META "save_meta_info"
+#define REC_VOLUMES "volumes"
+#define REC_SKIP "skip"
+#define REC_NAME "name"
+#define REC_DESCRIPTION "description"
+#define PROC_FLIP_BSCANS "flip_bscans"
+#define PROC_BITSHIFT "bitshift"
+#define PROC_MIN "min"
+#define PROC_MAX "max"
+#define PROC_LOG "log"
+#define PROC_COEFF "coeff"
+#define PROC_ADDEND "addend"
+#define PROC_RESAMPLING "resampling"
+#define PROC_RESAMPLING_INTERPOLATION "resampling_interpolation"
+#define PROC_RESAMPLING_C0 "resampling_c0"
+#define PROC_RESAMPLING_C1 "resampling_c1"
+#define PROC_RESAMPLING_C2 "resampling_c2"
+#define PROC_RESAMPLING_C3 "resampling_c3"
+#define PROC_DISPERSION_COMPENSATION "dispersion_compensation"
+#define PROC_DISPERSION_COMPENSATION_D0 "dispersion_compensation_d0"
+#define PROC_DISPERSION_COMPENSATION_D1 "dispersion_compensation_d1"
+#define PROC_DISPERSION_COMPENSATION_D2 "dispersion_compensation_d2"
+#define PROC_DISPERSION_COMPENSATION_D3 "dispersion_compensation_d3"
+#define PROC_WINDOWING "windowing"
+#define PROC_WINDOWING_TYPE "window_type"
+#define PROC_WINDOWING_FILL_FACTOR "window_fill_factor"
+#define PROC_WINDOWING_CENTER_POSITION "window_center_position"
[…]
