+++ octproz_project/octproz/src/cuda_code.cu
@@ -1,0 +1,1206 @@
+/**
+**  This file is part of OCTproZ.
+**  OCTproZ is an open source software for processig of optical
+**  coherence tomography (OCT) raw data.
+**  Copyright (C) 2019-2020 Miroslav Zabic
+**
+**  OCTproZ is free software: you can redistribute it and/or modify
+**  it under the terms of the GNU General Public License as published by
+**  the Free Software Foundation, either version 3 of the License, or
+**  (at your option) any later version.
+**
+**  This program is distributed in the hope that it will be useful,
+**  but WITHOUT ANY WARRANTY; without even the implied warranty of
+**  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+**  GNU General Public License for more details.
+**
+**  You should have received a copy of the GNU General Public License
+**  along with this program. If not, see http://www.gnu.org/licenses/.
+**
+****
+** Author:	Miroslav Zabic
+** Contact:	zabic
+**			at
+**			iqo.uni-hannover.de
+****
+**/
+
+#ifndef CUDA_CODE_CU
+#define CUDA_CODE_CU
+
+#include "kernels.h"
+
+int blockSize;
+int gridSize;
+
+surface<void, cudaSurfaceType3D> surfaceWrite;
+
+cudaStream_t copyStream;
+cudaStream_t copyStreamD2H;
+cudaStream_t streamingStreamD2H;
+cudaStream_t processStream;
+cudaStream_t displayStream;
+cudaStream_t userRequestStream;
+
+cudaEvent_t processEvent;
+
+cudaGraphicsResource* cuBufHandleBscan = NULL;
+cudaGraphicsResource* cuBufHandleEnFaceView = NULL;
+cudaGraphicsResource* cuBufHandleVolumeView = NULL;
+
+void* d_inputBuffer1;
+void* d_inputBuffer2;
+void* d_outputBuffer;
+
+
+void* host_buffer1 = NULL;
+void* host_buffer2 = NULL;
+void* host_RecordBuffer = NULL;
+void* host_streamingBuffer1;
+void* host_streamingBuffer2;
+
+cufftComplex* d_inputLinearized;
+float* d_windowCurve= NULL;
+float* d_resampleCurve = NULL;
+float* d_dispersionCurve = NULL;
+float* d_sinusoidalResampleCurve = NULL;
+cufftComplex* d_phaseCartesian = NULL;
+unsigned int bufferNumber = 0;
+unsigned int bufferNumberInVolume = 0;
+unsigned int streamingBufferNumber = 0;
+
+
+cufftComplex* d_fftBuffer = NULL;
+cufftHandle d_plan;
+cufftComplex* d_meanALine = NULL;
+
+bool cudaInitialized = false;
+bool saveToDisk = false;
+
+size_t signalLength = 0;
+size_t ascansPerBscan = 0;
+size_t bscansPerBuffer = 0;
+size_t samplesPerBuffer = 0;
+size_t samplesPerVolume = 0;
+size_t buffersPerVolume = 0;
+size_t bytesPerSample = 0;
+
+float* d_processedBuffer = NULL;
+float* d_sinusoidalScanTmpBuffer = NULL;
+OctAlgorithmParameters* params = NULL;
+
+bool firstRun = true;
+unsigned int processedBuffers;
+unsigned int streamedBuffers;
+
+bool fixedPatternNoiseDetermined = false;
+
+
+
[…]

[…]
