### 2020-10-05 #2580 @kthyng
Hi all! I am the rotating associate editor in chief for the week and want to check in here since it's been awhile since there has been activity on this review. It looks like we may have a complete review from @brandondube, no review from @jrasakanthan, and partial review from @jdavidli. I think we can take a step forward if @brandondube can verify if that review is finished, get a review started from @jrasakanthan, and see what the next step is for @jdavidli. Thanks all!
