### 2020-10-10 #2580 @jdavidli
@spectralcode I really like the new page you have on performance. It's super helpful and I was able to replicate your performance. Same hardware as previously mentioned.
Volumes/second: 38
Buffers/second: 38
B-scans/second: 9840
A-scans/second: 5.038e6
Buffer size: 256
Data throughput: 9840

I just have a few minor suggestions for the paper itself. In the third line of the introduction, it should say "by combining a reference beam." In section two, towards the end, where you talk about physical or virtual OCT systems, I think it would be useful to add a sentence explaining what you mean by virtual OCT systems. This also relates to my last concern. At the beginning of section three, you say raw data from the OCT system is transferred to RAM. I might've missed this, but how would is that simulated in a virtual system? In a physical OCT system, you would need to acquire data using some kind of acquisition card like you mentioned in section five or the processing page, and then transfer to RAM.
