### 2020-10-15 0dadde3
minor changes in paper.md
