### 2020-10-15 5fddf47
minor changes and fixed typo in paper.md
