### 2020-10-15 #2580 @spectralcode
@jdavidli thank you very much! Your performance test results look great!

> In the third line of the introduction, it should say "by combining a reference beam."

→ Thanks! I have changed it.

> In section two, towards the end, where you talk about physical or virtual OCT systems, I think it would be useful to add a sentence explaining what you mean by virtual OCT systems.

→ Thank you very much, you are right about this! In order not to make the paper any longer (the paper length is already slightly above the required length of [250 - 1000 words](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)), I rephrased the sentence so that the word "virtual" is no longer used. I hope this avoids any confusion regarding the virtual OCT system as it is not mentioned anymore. Please let me know what you think.

> raw data from the OCT system is transferred to RAM. I might've missed this, but how would is that simulated in a virtual system?

→ The plugin _Virtual OCT System_ reads raw data from the hard disk and transfers it to RAM. In an actual OCT system there would be some kind of acquisistion card (just like you mentioned) that does not need to store the acquired data on the hard disk. The data can be transferred directly to the RAM.
