### 2020-10-16 #2580 @arfon
@spectralcode - thanks for the updates.

@jdavidli - with these latest changes, are you ready to check off the remaining items in your reviewer checklist above?
