### 2020-10-26 #2580 @arfon
@spectralcode - could you please merge this PR which implements a few small fixes to your paper: https://github.com/spectralcode/OCTproZ/pull/5

After this could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
