### 2020-10-28 063031f
Processing pipeline figure for paper redesigned
