### 2020-10-28 3ef2d5d
minor change in paper (text below image shortened to avoid spreading the table at the bottom of the page on to the next page)
