### 2020-10-28 d958dbb
minor change in paper (text below image shortened to avoid spreading the table at the bottom of the page on to the next page)
