+++ paper/paper.md
@@ -42,7 +42,7 @@ OCTproZ performs live signal processing and visualization of OCT data. It is wri
  ![Screenshot of OCTproZ v1.0.0 Processing settings visible in the left panel can be changed before processing is started or while processing is in progress. Processed data is live visualized in 2D as cross sectional images (B-scan and en face view) and in 3D as interactive volume rendering. The live view shows a piece of wood with a couple layers of tape and a laser burned hole. \label{fig:screenshot}](figures/20191122_screenshot3d.png)
 
 
-The software can be separated into three parts: main application, development kit (DevKit) and plug-ins. The main application, OCTproZ itself, contains the logic for the GUI, processing and visualization. The DevKit, which is implemented as static library, provides the necessary interface for plug-in development. Plug-ins can be one of two kinds: “Acquisition Systems” or “Extensions”. The former represent OCT systems and provide raw data to OCTproZ, the later are software modules that can extend the functionality of an OCT system (e.g., software control of a liquid lens) or provide additional custom defined post processing steps. Both, Acquisition Systems and Extensions, are dynamic libraries that can be loaded into OCTproZ during runtime.  
+The software can be separated into three parts: main application, development kit (DevKit) and plug-ins. The main application, OCTproZ itself, contains the logic for the GUI, processing and visualization. The DevKit, which is implemented as static library, provides the necessary interface for plug-in development. Plug-ins can be one of two kinds: “Acquisition Systems” or “Extensions”. The former represent OCT systems and provide raw data to OCTproZ, the later are software modules that can extend the functionality of an OCT system (e.g., software control of a liquid lens) or provide additional custom defined post processing steps. All plug-ins are dynamic libraries that can be loaded into OCTproZ during runtime. 
 
 # 3. Processing Pipeline
 
@@ -50,7 +50,7 @@ Raw data, i.e. acquired spectral fringe pattern, from the OCT system is transfer
 
  ![Processing pipeline of OCTproZ v1.2.0. Each box inside "OCTproZ GPU Processing" represents a CUDA kernel. Some processing steps are combinend into a single kernel (e.g. k-linearization, dispersion compensation and windowing) to enhance processing performance. \label{fig:processing}](figures/processing_pipeline_linear_v2.png) 
 
-The processing pipeline consists of data conversion, k-linearization, numerical dispersion compensation, windowing, fast Fourier transform, fixed-pattern noise removal, truncate, logarithm, backward scan correction, sinusoidal scan correction and visualization. A detailed description of each processing step can be found in the user manual. Here we just want to mention that the fixed-pattern noise removal step is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster [@pilia2018raycaster].
+The processing pipeline consists of data conversion, k-linearization, numerical dispersion compensation, windowing, fast Fourier transform, fixed-pattern noise removal, truncate, logarithm, backward scan correction, sinusoidal scan correction and visualization. A detailed description of each processing step can be found in the user manual. Here we just want to mention that the implementation of the fixed-pattern noise removal is based on a publication by Moon et al. [@moon2010reference] and the volume viewer is based on source code from an open source raycaster [@pilia2018raycaster].
 In order to avoid unnecessary data transfer to host memory, CUDA-OpenGL interoperability is used which allows the processed data to remain in GPU memory for visualization. 
 
 # 4. Processing Performance 
