### 2020-10-28 #2580 @spectralcode
@arfon Thank you for the fixes. I merged your pull request!
Here is the DOI of the Zenodo archive: https://doi.org/10.5281/zenodo.4148992

The version number of the most recent release is v1.2.1

Please note: I have made some minor changes to the paper. One change that is noticable is figure 2, which I redesigned to make the text in the figure easier to read. I hope this is fine for everybody, please let me know if this is not the case.
