### 2020-10-29 #2580 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
